I have added the Java programs for sending Email and user verification.
Since I ran into a problem earlier today with Java not wanting to run my installed copy of ACL2,  I have added a config folder with a file called ACL2 path. I have it set at simply "acl2." I will need to change it to my saved acl2 path on my machine in order for it to work. This is the lazy solution, and I'm not wanting to mess with my ACL2 setup at this point.